# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/short-links/{id}':
    put:
      summary: '/**'
      operationId: ''
      description: 'Update short link by id'
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"id\": 1,\n\"link\": \"https://example.com\",\n\"use_js_redirect\": \"false\",\n\"user_id\": 1,\n\"short_link\": https://example.com/a3ew\n}"
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Unauthenticated
                properties:
                  error:
                    type: string
                    example: Unauthenticated
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Unauthorized
                properties:
                  error:
                    type: string
                    example: Unauthorized
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Link not found'
                properties:
                  error:
                    type: string
                    example: 'Link not found'
      tags:
        - Endpoints
      security: []
    delete:
      summary: 'Create new short link'
      operationId: createNewShortLink
      description: ''
      parameters: []
      responses:
        204:
          description: ''
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Unauthenticated
                properties:
                  error:
                    type: string
                    example: Unauthenticated
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Unauthorized
                properties:
                  error:
                    type: string
                    example: Unauthorized
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Link not found'
                properties:
                  error:
                    type: string
                    example: 'Link not found'
      tags:
        - Endpoints
      security: []
    get:
      summary: 'Get short link by id'
      operationId: getShortLinkById
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"id\": 1,\n\"link\": \"https://example.com\",\n\"use_js_redirect\": \"false\",\n\"user_id\": 1,\n\"short_link\": https://example.com/a3ew\n}"
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Unauthenticated
                properties:
                  error:
                    type: string
                    example: Unauthenticated
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Unauthorized
                properties:
                  error:
                    type: string
                    example: Unauthorized
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Link not found'
                properties:
                  error:
                    type: string
                    example: 'Link not found'
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the short link.'
        example: nisi
        required: true
        schema:
          type: string
...